+++ Example-iOS/TestCollectionViewAdapter/MainViewController.swift
@@ -54,7 +54,7 @@ class MainViewController: UIViewController {
 }
 
 class LineCell: UICollectionViewCell, CollectionViewAdapterCellProtocol {
-    static var SpanSize: Int = 0
+    static var SpanSize: CGFloat = 0
     
     var actionClosure: ((_ name: String, _ object: Any?) -> Void)?
